// src/types.d.ts
export declare enum StorageType {
    LOCAL = "local",
    GCS = "gcs",
    S3 = "s3",
    B2 = "b2"
}
export declare type JSON = {
    [id: string]: string | number | boolean | string[] | number[] | boolean[] | {
        [id: string]: JSON;
    };
};
export interface IAdapterConfig {
    type: string;
    slug?: boolean;
    bucketName?: string;
}
export declare type GenericKey = number | string | boolean | number[] | string[] | boolean[];
export interface ConfigAmazonS3 extends IAdapterConfig {
    accessKeyId: string;
    secretAccessKey: string;
    region?: string;
    endpoint?: string;
    useDualstack?: boolean;
    maxRetries?: number;
    maxRedirects?: number;
    sslEnabled?: boolean;
    [id: string]: GenericKey;
}
export interface ConfigBackblazeB2 extends IAdapterConfig {
    applicationKeyId: string;
    applicationKey: string;
}